Add support for standard units in light presets

+++ src/Aeon.Environment/RoomLightPreset.cs
@@ -3,11 +3,11 @@
     public struct RoomLightPreset
     {
         public string Name;
-        public int ColdWhite;
-        public int WarmWhite;
-        public int Red;
+        public float ColdWhite;
+        public float WarmWhite;
+        public float Red;
 
-        public RoomLightPreset(string name, int coldWhite, int warmWhite, int red)
+        public RoomLightPreset(string name, float coldWhite, float warmWhite, float red)
         {
             Name = name;
             ColdWhite = coldWhite;
